Set 2 Rebrand

+++ include/widgets/textbox.h
@@ -60,7 +60,10 @@ typedef struct TBFontConfig {
 typedef struct {
   widget widget;
   unsigned long flags;
-  short cursor;
+  // Action purpose: every producer computes this as an int from
+  // g_utf8_strlen; a narrower type wraps negative past SHRT_MAX characters and
+  // g_utf8_offset_to_pointer then walks backwards out of the allocation.
+  int cursor;
   char *text;
   char *placeholder;
   int show_placeholder;
